Integrate nifti test file creation into test running
